files in bridge separated
